\name{MergerAnalysis-package}
\alias{MergerAnalysis-package}
\alias{MergerAnalysis}
\docType{package}
\title{
\packageTitle{MergerAnalysis}
}
\description{
\packageDescription{MergerAnalysis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MergerAnalysis}
\packageIndices{MergerAnalysis}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MergerAnalysis}

Maintainer: \packageMaintainer{MergerAnalysis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
